tratando possiveis erros e adicionando loading

+++ app/src/main/java/com/jisellemartins/gallery/model/Data.kt
@@ -3,5 +3,5 @@ package com.jisellemartins.gallery.model
 class Data (
     val data: List<Image>,
     val sucess: Boolean,
-    val status: Float
+    val status: Int
 )
